@@ -12,15 +12,15 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "7.2.4",
-    "@angular/common": "7.2.4",
-    "@angular/compiler": "7.2.4",
-    "@angular/core": "7.2.4",
-    "@angular/forms": "7.2.4",
-    "@angular/http": "7.2.4",
-    "@angular/platform-browser": "7.2.4",
-    "@angular/platform-browser-dynamic": "7.2.4",
-    "@angular/router": "7.2.4",
+    "@angular/animations": "8.0.0-beta.3",
+    "@angular/common": "8.0.0-beta.3",
+    "@angular/compiler": "8.0.0-beta.3",
+    "@angular/core": "8.0.0-beta.3",
+    "@angular/forms": "8.0.0-beta.3",
+    "@angular/http": "8.0.0-beta.3",
+    "@angular/platform-browser": "8.0.0-beta.3",
+    "@angular/platform-browser-dynamic": "8.0.0-beta.3",
+    "@angular/router": "8.0.0-beta.3",
     "@ngrx/effects": "^7.2.0",
     "@ngrx/router-store": "^7.2.0",
     "@ngrx/store": "^7.2.0",
@@ -33,9 +33,9 @@
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^0.13.0",
-    "@angular/cli": "7.3.0",
-    "@angular/compiler-cli": "7.2.4",
-    "@angular/language-service": "7.2.4",
+    "@angular/cli": "7.3.1",
+    "@angular/compiler-cli": "8.0.0-beta.3",
+    "@angular/language-service": "8.0.0-beta.3",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "~6.0.60",
