@@ -1,75 +1,72 @@
-import { Injectable } from '@angular/core';
-import { Effect, Actions, ofType } from '@ngrx/effects';
-import { Observable, from, pipe } from 'rxjs';
-import { Action } from '@ngrx/store';
-import { map, switchMap, mergeMap, tap } from 'rxjs/operators';
+import {Injectable} from '@angular/core';
+import {Actions, Effect, ofType} from '@ngrx/effects';
+import {Router} from '@angular/router';
+import {map, tap, switchMap, mergeMap} from 'rxjs/operators';
+import { from } from 'rxjs';
 import * as firebase from 'firebase';
-import { Router } from '@angular/router';
 
 import * as AuthActions from './auth.actions';
 
 @Injectable()
 export class AuthEffects {
   @Effect()
-  authSignup: Observable<Action> = this.actions$.pipe(
-    ofType(AuthActions.TRY_SIGNUP),
-    map((actions: AuthActions.TrySignup) => {
-      return actions.payload;
-    }),
-    switchMap((authData: {username: string, password: string}) => {
-      return from(firebase.auth().createUserWithEmailAndPassword(authData.username, authData.password));
-    }),
-    switchMap(() => {
-      return from(firebase.auth().currentUser.getIdToken());
-    }),
-    mergeMap((token: string) => {
-      return [
-        {
-          type: AuthActions.SIGNUP
-        },
-        {
-          type: AuthActions.SET_TOKEN,
-          payload: token
-        }
-      ];
-    })
-  );
+  authSignup = this.actions$
+    .pipe(
+      ofType(AuthActions.TRY_SIGNUP),
+      map((action: AuthActions.TrySignup) => {
+        return action.payload;
+      })
+      , switchMap((authData: { username: string, password: string }) => {
+        return from(firebase.auth().createUserWithEmailAndPassword(authData.username, authData.password));
+      })
+      , switchMap(() => {
+        return from(firebase.auth().currentUser.getIdToken());
+      })
+      , mergeMap((token: string) => {
+        return [
+          {
+            type: AuthActions.SIGNUP
+          },
+          {
+            type: AuthActions.SET_TOKEN,
+            payload: token
+          }
+        ];
+      }));
 
   @Effect()
-  authSignin = this.actions$.pipe(
-    ofType(AuthActions.TRY_SIGNIN),
-    map((actions: AuthActions.TrySingin) => {
-      return actions.payload;
-    }),
-    switchMap((authData: {username: string, password: string}) => {
-      return from(firebase.auth().signInWithEmailAndPassword(authData.username, authData.password));
-    }),
-    switchMap(() => {
-      return from(firebase.auth().currentUser.getIdToken());
-    }),
-    mergeMap((token: string) => {
-      return [
-        {
-          type: AuthActions.SIGNIN
-        },
-        {
-          type: AuthActions.SET_TOKEN,
-          payload: token
-        }
-      ];
-    })
-  );
+  authSignin = this.actions$
+    .pipe(
+      ofType(AuthActions.TRY_SIGNIN),
+      map((action: AuthActions.TrySignup) => {
+        return action.payload;
+      })
+      , switchMap((authData: { username: string, password: string }) => {
+        return from(firebase.auth().signInWithEmailAndPassword(authData.username, authData.password));
+      })
+      , switchMap(() => {
+        return from(firebase.auth().currentUser.getIdToken());
+      })
+      , mergeMap((token: string) => {
+        this.router.navigate(['/']);
+        return [
+          {
+            type: AuthActions.SIGNIN
+          },
+          {
+            type: AuthActions.SET_TOKEN,
+            payload: token
+          }
+        ];
+      }));
 
   @Effect({dispatch: false})
-  authLogout = this.actions$.pipe(
-    ofType(AuthActions.LOGOUT),
-    tap(() => {
+  authLogout = this.actions$
+    .pipe(
+      ofType(AuthActions.LOGOUT),
+      tap(() => {
       this.router.navigate(['/']);
-    })
-  );
+    }));
 
-  constructor(
-    private actions$: Actions,
-    private router: Router
-  ) {}
+  constructor(private actions$: Actions, private router: Router) {}
 }
