@@ -2,24 +2,23 @@ import { HttpEvent, HttpHandler, HttpInterceptor, HttpRequest } from '@angular/c
 import { Observable } from 'rxjs';
 import { Injectable } from '@angular/core';
 import { Store } from '@ngrx/store';
-import { switchMap, take } from 'rxjs/operators';
+import { take, switchMap } from 'rxjs/operators';
 
-import * as fromApp from '../app-ngrx-store/app.reducers';
-import * as fromAuth from '../auth/auth-ngrx-store/auth.reducers';
+import * as fromAppReducers from '../app-ngrx-store/app.reducers';
+import * as fromAuthReducers from '../auth/auth-ngrx-store/auth.reducers';
 
 @Injectable()
 export class AuthInterceptor implements HttpInterceptor {
-  constructor(private store: Store<fromApp.IAppState>) {}
+  constructor(private store: Store<fromAppReducers.IAppState>) {}
 
   intercept(req: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
     console.log('Intercepted!', req);
 
-    return this.store.select('auth').pipe(
-      take(1),
-      switchMap((authState: fromAuth.IState) => {
-        const copiedReq = req.clone({params: req.params.set('auth', authState.token)});
-        return next.handle(copiedReq);
-      })
-    );
+    return this.store.select('auth')
+      .pipe(take(1),
+        switchMap((authState: fromAuthReducers.IState) => {
+          const copiedReq = req.clone({params: req.params.set('auth', authState.token)});
+          return next.handle(copiedReq);
+        }));
   }
 }
