@@ -7,26 +7,27 @@ export const SIGNIN = 'SIGNIN';
 export const LOGOUT = 'LOGOUT';
 export const SET_TOKEN = 'SET_TOKEN';
 
-export class TrySignup {
+export class TrySignup implements Action {
   readonly type = TRY_SIGNUP;
 
   constructor(public payload: {username: string, password: string}) {}
 }
-export class SignUp implements Action {
-  readonly type = SIGNUP;
-}
 
-export class TrySingin {
-  readonly type = SIGNIN;
+export class TrySignin implements Action {
+  readonly type = TRY_SIGNIN;
 
   constructor(public payload: {username: string, password: string}) {}
 }
 
-export class SignIn implements Action {
+export class Signup implements Action {
+  readonly type = SIGNUP;
+}
+
+export class Signin implements Action {
   readonly type = SIGNIN;
 }
 
-export class LogOut implements Action {
+export class Logout implements Action {
   readonly type = LOGOUT;
 }
 
@@ -37,9 +38,9 @@ export class SetToken implements Action {
 }
 
 export type AuthActions =
-  SignUp
-| SignIn
-| LogOut
+  Signup
+| Signin
+| Logout
 | SetToken
 | TrySignup
-| TrySingin;
+| TrySignin;
