@@ -1,7 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { NgForm } from '@angular/forms';
 import { Store } from '@ngrx/store';
-import { Router } from '@angular/router';
 
 import * as fromApp from '../../app-ngrx-store/app.reducers';
 import * as AuthActions from '../auth-ngrx-store/auth.actions';
@@ -14,8 +13,7 @@ import * as AuthActions from '../auth-ngrx-store/auth.actions';
 export class SigninComponent implements OnInit {
 
   constructor(
-    private store: Store<fromApp.IAppState>,
-    private router: Router
+    private store: Store<fromApp.IAppState>
   ) { }
 
   ngOnInit() {
@@ -25,8 +23,7 @@ export class SigninComponent implements OnInit {
     const email = form.value.email;
     const password = form.value.password;
 
-    this.store.dispatch(new AuthActions.TrySingin({username: email, password: password}));
-    this.router.navigate(['/']);
+    this.store.dispatch(new AuthActions.TrySignin({username: email, password: password}));
   }
 
 }
