@@ -7,7 +7,7 @@ import { Store } from '@ngrx/store';
 import * as RecipeActions from '../recipes-ngrx-store/recipe.actions';
 import { RecipeModel } from '../recipe.model';
 import { secrets } from '../../../secrets';
-import * as fromAppReducers from '../../app-ngrx-store/app.reducers';
+import * as fromRecipeReducers from '../recipes-ngrx-store/recipe.reducers';
 
 @Injectable()
 export class RecipeEffects {
@@ -18,42 +18,37 @@ export class RecipeEffects {
     .pipe(
       ofType(RecipeActions.FETCH_RECIPES),
       switchMap((action: RecipeActions.FetchRecipes) => {
-        return this.httpClient.get<RecipeModel[]>(this.apiUrl, {
-          observe: 'body',
-          responseType: 'json'
-        });
-      }),
-      map(
-        (recipes) => {
-          console.log(recipes);
-          for (const recipe of recipes) {
-            if (!recipe['ingredients']) {
-              recipe['ingredients'] = [];
-            }
+      return this.httpClient.get<RecipeModel[]>(this.apiUrl, {
+        observe: 'body',
+        responseType: 'json'
+      });
+    }), map(
+      (recipes) => {
+        console.log(recipes);
+        for (const recipe of recipes) {
+          if (!recipe['ingredients']) {
+            recipe['ingredients'] = [];
           }
-          return {
-            type: RecipeActions.SET_RECIPES,
-            payload: recipes
-          };
         }
-      )
-    );
+        return {
+          type: RecipeActions.SET_RECIPES,
+          payload: recipes
+        };
+      }
+    ));
 
-  @Effect({ dispatch: false })
+  @Effect({dispatch: false})
   recipeStore = this.actions$
-      .pipe(
-        ofType(RecipeActions.STORE_RECIPES),
-        withLatestFrom(this.store.select('recipes')),
-        switchMap(([action, state]) => {
-          const req = new HttpRequest('PUT', this.apiUrl, state.recipes, { reportProgress: true });
+    .pipe(
+      ofType(RecipeActions.STORE_RECIPES),
+      withLatestFrom(this.store.select('recipes')),
+      switchMap(([action, state]) => {
+        const req = new HttpRequest('PUT', this.apiUrl, state.recipes, {reportProgress: true});
+        return this.httpClient.request(req);
+      }));
 
-          return this.httpClient.request(req);
-        })
-      );
-
-  constructor(
-    private actions$: Actions,
-    private httpClient: HttpClient,
-    private store: Store<fromAppReducers.IAppState>
-  ) { }
+  constructor(private actions$: Actions,
+              private httpClient: HttpClient,
+              private store: Store<fromRecipeReducers.IFeatureState>) {
+  }
 }
