@@ -31,33 +31,31 @@ const initialState: IState = {
 
 export function recipeReducer(state = initialState, action: RecipeActions.RecipeActions) {
   switch (action.type) {
-    case RecipeActions.SET_RECIPES:
+    case (RecipeActions.SET_RECIPES):
       return {
         ...state,
         recipes: [...action.payload]
       };
-    case RecipeActions.ADD_RECIPE:
+    case (RecipeActions.ADD_RECIPE):
       return {
         ...state,
         recipes: [...state.recipes, action.payload]
       };
-    case RecipeActions.UPDATE_RECIPE:
+    case (RecipeActions.UPDATE_RECIPE):
       const recipe = state.recipes[action.payload.index];
       const updatedRecipe = {
         ...recipe,
         ...action.payload.updatedRecipe
       };
-      const _recipes = [...state.recipes];
-      _recipes[action.payload.index] = updatedRecipe;
-
+      const recipes = [...state.recipes];
+      recipes[action.payload.index] = updatedRecipe;
       return {
         ...state,
-        recipe: _recipes
+        recipes: recipes
       };
-    case RecipeActions.DELETE_RECIPE:
+    case (RecipeActions.DELETE_RECIPE):
       const oldRecipes = [...state.recipes];
       oldRecipes.splice(action.payload, 1);
-
       return {
         ...state,
         recipes: oldRecipes
